+++ src/test/java/com/cognite/client/EventsIntegrationTest.java
@@ -246,12 +246,10 @@ class EventsIntegrationTest {
                     Duration.between(startInstant, Instant.now()));
 
             LOG.info(loggingPrefix + "Start retrieving events.");
-            List<Item> eventItems = new ArrayList<>();
-            listEventsResults.stream()
-                    .map(event -> Item.newBuilder()
-                            .setExternalId(event.getExternalId())
-                            .build())
-                    .forEach(item -> eventItems.add(item));
+            List<String> eventExternalIds = listEventsResults.stream()
+                    .map(Event::getExternalId)
+                    .collect(Collectors.toList());
+
 
             List<Event> retrievedEvents = client.events().retrieve(eventItems);
             LOG.info(loggingPrefix + "Finished retrieving events. Duration: {}",